# swagger.json
{
  "openapi": "3.1.0",
  "servers": [
    {
      "url": "{scheme}://{host}/api/v1",
      "variables": {
        "scheme": {
          "enum": [
            "https",
            "http"
          ],
          "default": "https"
        },
        "host": {
          "default": "localhost:3000"
        }
      }
    },
    {
      "url": "/api/v1"
    }
  ],
  "info": {
    "version": "0.0.1",
    "title": "mycelia-api",
    "description": "",
    "license": {
      "name": "Apache-2.0"
    }
  },
  "paths": {
...
    "/embeddings/batch-update": {
      "post": {
        "tags": [
          "Embeddings"
        ],
        "parameters": [],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {}
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "embeddings",
          "batch_update",
          "triggerBatchUpdate"
        ],
        "x-samchon-controller": "EmbeddingController"
      }
    },
    "/embeddings/progress": {
      "get": {
        "tags": [
          "Embeddings"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "embeddings",
          "progress",
          "getBatchProgress"
        ],
        "x-samchon-controller": "EmbeddingController"
      }
    },
...
    "/favorites": {
      "get": {
        "tags": [
          "Favorites"
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "favorites",
          "getUserFavorites"
        ],
        "x-samchon-controller": "FavoriteController"
      }
    },
    "/favorites/{feedItemId}/is-favorited": {
      "get": {
        "tags": [
          "Favorites"
        ],
        "parameters": [
          {
            "name": "feedItemId",
            "in": "path",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "favorites",
          "is_favorited",
          "checkFavorite"
        ],
        "x-samchon-controller": "FavoriteController"
      }
    },
    "/favorites/{feedItemId}": {
      "post": {
        "tags": [
          "Favorites"
        ],
        "parameters": [
          {
            "name": "feedItemId",
            "in": "path",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "favorites",
          "favoriteItem"
        ],
        "x-samchon-controller": "FavoriteController"
      },
      "delete": {
        "tags": [
          "Favorites"
        ],
        "parameters": [
          {
            "name": "feedItemId",
            "in": "path",
            "schema": {
              "type": "number"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          }
        },
        "security": [
          {
            "bearer": []
          }
        ],
        "x-samchon-accessor": [
          "favorites",
          "unfavoriteItem"
        ],
        "x-samchon-controller": "FavoriteController"
      }
    },
...
  },
  "components": {
...
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
...
}